Backend/requirements.txt
pandas==2.2.2
smbus==1.1.post2
requests==2.32.3
python-dotenv==1.0.1